+++ opentorsion/gear_element.py
