+++ opentorsion/shaft_element.py
